Refactor DB support with API types
