+++ db/src/main/scala/squeryl/BlockBondsDbApiImpl.scala
@@ -1,18 +1,19 @@
-package db
+package squeryl
 
 import cats.effect.Sync
 import cats.syntax.all.*
 import sdk.CustomTypeMode.*
 import sdk.RhonixNodeDb.blockBondsTable
+import sdk.api.BlockBonds
 import sdk.db.DbSession.withSessionF
-import sdk.db.{BlockBonds, BlockBondsDbApi, DbSession}
+import sdk.db.{BlockBondsDbApi, BlockBondsTable, DbSession}
 
 class BlockBondsDbApiImpl[F[_]: Sync: DbSession] extends BlockBondsDbApi[F] {
   override def insert(blockBonds: BlockBonds): F[BlockBonds] =
-    withSessionF(blockBondsTable.insert(BlockBonds.toDb(blockBonds))).map(BlockBonds.fromDb)
+    withSessionF(blockBondsTable.insert(BlockBondsTable.toDb(blockBonds))).map(BlockBondsTable.fromDb)
 
   override def getByBlock(blockId: Long): F[Seq[BlockBonds]] =
-    withSessionF(blockBondsTable.where(_.blockId === blockId).map(BlockBonds.fromDb).toSeq)
+    withSessionF(blockBondsTable.where(_.blockId === blockId).map(BlockBondsTable.fromDb).toSeq)
 }
 
 object BlockBondsDbApiImpl {
