+++ db/src/main/scala/squeryl/BlockDeploysDbApiImpl.scala
@@ -1,18 +1,19 @@
-package db
+package squeryl
 
 import cats.effect.Sync
 import cats.syntax.all.*
 import sdk.CustomTypeMode.*
 import sdk.RhonixNodeDb.blockDeploysTable
+import sdk.api.BlockDeploys
 import sdk.db.DbSession.withSessionF
-import sdk.db.{BlockDeploys, BlockDeploysDbApi, DbSession}
+import sdk.db.{BlockDeploysDbApi, BlockDeploysTable, DbSession}
 
 class BlockDeploysDbApiImpl[F[_]: Sync: DbSession] extends BlockDeploysDbApi[F] {
   override def insert(blockDeploys: BlockDeploys): F[BlockDeploys] =
-    withSessionF(blockDeploysTable.insert(BlockDeploys.toDb(blockDeploys))).map(BlockDeploys.fromDb)
+    withSessionF(blockDeploysTable.insert(BlockDeploysTable.toDb(blockDeploys))).map(BlockDeploysTable.fromDb)
 
   override def getByBlock(blockId: Long): F[Seq[BlockDeploys]] = withSessionF(
-    blockDeploysTable.where(_.blockId === blockId).map(BlockDeploys.fromDb).toSeq,
+    blockDeploysTable.where(_.blockId === blockId).map(BlockDeploysTable.fromDb).toSeq,
   )
 }
