+++ db/src/main/scala/squeryl/BondDbApiImpl.scala
@@ -1,25 +1,26 @@
-package db
+package squeryl
 
 import cats.data.OptionT
 import cats.effect.Sync
 import cats.syntax.all.*
 import sdk.CustomTypeMode.*
 import sdk.RhonixNodeDb.{bondTable, validatorTable}
+import sdk.api.Bond
 import sdk.db.DbSession.withSessionF
-import sdk.db.{Bond, BondDbApi, DbSession}
+import sdk.db.{BondDbApi, BondTable, DbSession}
 
 class BondDbApiImpl[F[_]: Sync: DbSession] extends BondDbApi[F] {
   override def insert(bond: Bond, validatorId: Long): F[Long] =
-    withSessionF(bondTable.insert(Bond.toDb(0L, bond, validatorId))).map(_.id)
+    withSessionF(bondTable.insert(BondTable.toDb(0L, bond, validatorId))).map(_.id)
 
   override def update(id: Long, bond: Bond, validatorId: Long): F[Unit] =
-    withSessionF(bondTable.update(Bond.toDb(id, bond, validatorId)))
+    withSessionF(bondTable.update(BondTable.toDb(id, bond, validatorId)))
 
   override def getById(id: Long): F[Option[Bond]] =
     (for {
       bond      <- OptionT(withSessionF(bondTable.where(_.id === id).headOption))
       validator <- OptionT(withSessionF(validatorTable.where(_.id === bond.validatorId).headOption))
-    } yield Bond.fromDb(bond, validator)).value
+    } yield BondTable.fromDb(bond, validator)).value
 }
 
 object BondDbApiImpl {
