+++ db/src/main/scala/squeryl/DeployDbApiImpl.scala
@@ -1,24 +1,25 @@
-package db
+package squeryl
 
 import cats.effect.Sync
 import cats.syntax.all.*
 import sdk.CustomTypeMode.*
 import sdk.RhonixNodeDb.deployTable
+import sdk.api.Deploy
 import sdk.db.*
 import sdk.db.DbSession.withSessionF
 
 class DeployDbApiImpl[F[_]: Sync: DbSession] extends DeployDbApi[F] {
   override def insert(deploy: Deploy): F[Long] =
-    withSessionF(deployTable.insert(Deploy.toDb(0L, deploy))).map(_.id)
+    withSessionF(deployTable.insert(DeployTable.toDb(0L, deploy))).map(_.id)
 
   override def update(id: Long, deploy: Deploy): F[Unit] =
-    withSessionF(deployTable.update(Deploy.toDb(id, deploy)))
+    withSessionF(deployTable.update(DeployTable.toDb(id, deploy)))
 
   override def getById(id: Long): F[Option[Deploy]] =
-    withSessionF(deployTable.where(_.id === id).headOption.map(d => Deploy.fromDb(d)))
+    withSessionF(deployTable.where(_.id === id).headOption.map(d => DeployTable.fromDb(d)))
 
   override def getByHash(hash: Array[Byte]): F[Option[Deploy]] =
-    withSessionF(deployTable.where(_.hash === hash).headOption.map(d => Deploy.fromDb(d)))
+    withSessionF(deployTable.where(_.hash === hash).headOption.map(d => DeployTable.fromDb(d)))
 }
 
 object DeployDbApiImpl {
