+++ db/src/main/scala/squeryl/ExternalApiImpl.scala
@@ -1,0 +1,16 @@
+package squeryl
+
+import cats.effect.Sync
+import sdk.api.{Block, Deploy, ExternalApi}
+import sdk.db.DbSession.withSessionF
+import sdk.db.{DbSession, DeployTable}
+
+/**
+ * Implementation of the external API using `squeryl` library.
+ */
+class ExternalApiImpl[F[_]: Sync: DbSession] extends ExternalApi[F] {
+  override def getBlockById(blockId: Long): F[Option[Block]] = ???
+
+  override def getDeploysByBlockId(blockId: Long): F[Seq[Deploy]] =
+    withSessionF(DbQuery.getDeploysByBlockId(blockId).iterator.map(DeployTable.fromDb).toSeq)
+}
