+++ db/src/main/scala/squeryl/ValidatorDbApiImpl.scala
@@ -1,24 +1,25 @@
-package db
+package squeryl
 
 import cats.effect.Sync
 import cats.syntax.all.*
 import sdk.CustomTypeMode.*
 import sdk.RhonixNodeDb.validatorTable
+import sdk.api.Validator
 import sdk.db.DbSession.withSessionF
-import sdk.db.{DbSession, Validator, ValidatorDbApi}
+import sdk.db.{DbSession, ValidatorDbApi, ValidatorTable}
 
 class ValidatorDbApiImpl[F[_]: Sync: DbSession] extends ValidatorDbApi[F] {
   override def insert(validator: Validator): F[Long] =
-    withSessionF(validatorTable.insert(Validator.toDb(0L, validator))).map(_.id)
+    withSessionF(validatorTable.insert(ValidatorTable.toDb(0L, validator))).map(_.id)
 
   override def update(id: Long, validator: Validator): F[Unit] =
-    withSessionF(validatorTable.update(Validator.toDb(id, validator)))
+    withSessionF(validatorTable.update(ValidatorTable.toDb(id, validator)))
 
   override def getById(id: Long): F[Option[Validator]] =
-    withSessionF(validatorTable.where(_.id === id).headOption.map(v => Validator.fromDb(v)))
+    withSessionF(validatorTable.where(_.id === id).headOption.map(v => ValidatorTable.fromDb(v)))
 
   override def getByPublicKey(publicKey: Array[Byte]): F[Option[Validator]] =
-    withSessionF(validatorTable.where(_.publicKey === publicKey).headOption.map(v => Validator.fromDb(v)))
+    withSessionF(validatorTable.where(_.publicKey === publicKey).headOption.map(v => ValidatorTable.fromDb(v)))
 }
 
 object ValidatorDbApiImpl {
