+++ sdk/src/main/scala/sdk/api/ExternalApi.scala
@@ -1,0 +1,6 @@
+package sdk.api
+
+trait ExternalApi[F[_]] {
+  def getBlockById(blockId: Long): F[Option[Block]]
+  def getDeploysByBlockId(blockId: Long): F[Seq[Deploy]]
+}
