+++ sdk/src/main/scala/sdk/db/BlockDeploysTable.scala
@@ -1,16 +1,16 @@
 package sdk.db
 
+import sdk.api.BlockDeploys
+
 @SuppressWarnings(Array("org.wartremover.warts.FinalCaseClass"))
 case class BlockDeploysTable(blockId: Long, deployId: Long)
-
-final case class BlockDeploys(blockId: Long, deployId: Long)
 
 trait BlockDeploysDbApi[F[_]] {
   def insert(blockDeploys: BlockDeploys): F[BlockDeploys]
   def getByBlock(blockId: Long): F[Seq[BlockDeploys]]
 }
 
-object BlockDeploys {
+object BlockDeploysTable {
   def toDb(blockDeploys: BlockDeploys): BlockDeploysTable =
     BlockDeploysTable(blockDeploys.blockId, blockDeploys.deployId)
