+++ sdk/src/main/scala/sdk/db/BlockTable.scala
@@ -1,6 +1,7 @@
 package sdk.db
 
 import sdk.DbTable
+import sdk.api.{Block, Bond, Deploy, Validator}
 
 @SuppressWarnings(Array("org.wartremover.warts.FinalCaseClass"))
 case class BlockTable(
@@ -27,55 +28,7 @@ case class BlockTable(
   status: Int,
 ) extends DbTable
 
-final case class Block(
-  hash: Array[Byte],
-  sender: Validator,
-  version: Int,
-  shardId: String,
-  seqNum: Long,
-  number: Long,
-  justificationsHash: Array[Byte], // hash(Set[Justification]),
-  justifications: Set[Validator],
-  bondsHash: Array[Byte],          // hash(Map[Validator, Long]),
-  bonds: Set[Bond],
-
-  // Rholang (tuple space) state change
-  preStateHash: Array[Byte],  // hash(VM state)
-  postStateHash: Array[Byte], // hash(VM state)
-  deploysHash: Array[Byte],   // hash(Set[Deploy])
-  deploys: Set[Deploy],
-
-  // Block signature
-  signatureAlg: String,
-  signature: Array[Byte],
-
-  // Status
-  status: Int,
-) {
-  override def equals(obj: Any): Boolean = obj match {
-    case that: Block =>
-      this.hash.sameElements(that.hash) &&
-      this.sender == that.sender &&
-      this.version == that.version &&
-      this.shardId == that.shardId &&
-      this.seqNum == that.seqNum &&
-      this.number == that.number &&
-      this.justificationsHash.sameElements(that.justificationsHash) &&
-      this.justifications == that.justifications &&
-      this.bondsHash.sameElements(that.bondsHash) &&
-      this.bonds == that.bonds &&
-      this.preStateHash.sameElements(that.preStateHash) &&
-      this.postStateHash.sameElements(that.postStateHash) &&
-      this.deploysHash.sameElements(that.deploysHash) &&
-      this.deploys == this.deploys &&
-      this.signatureAlg == this.signatureAlg &&
-      this.signature.sameElements(that.signature) &&
-      this.status == that.status
-    case _           => false
-  }
-}
-
-object Block {
+object BlockTable {
   def toDb(id: Long, block: Block, senderId: Long): BlockTable = BlockTable(
     id,
     block.hash,
